INC1-201 fix: remove unused code in SignIn component and refactoring
