+++ src/app/auth/sign-in/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { SignIn } from '@/features/auth/ui/sign-in'
+import { SignIn } from '@/features/auth'
 
 export default function SignInPage() {
   return <SignIn />
