+++ src/app/page.tsx
@@ -1,3 +1,12 @@
+'use client'
+
+import { SignIn } from '@/features/auth/sign-in/ui/SignIn'
+
 export default function Home() {
-  return <div>Hello, TechnoWhales!</div>
+  return (
+    <>
+      <div>Hello, TechnoWhales!</div>
+      <SignIn />
+    </>
+  )
 }
