+++ src/features/auth/index.ts
@@ -1,1 +1,2 @@
 export { OAuth } from './oAuth/ui/OAuth'
+export { SignIn } from 'src/features/auth/sign-in/ui/SignIn'
