--- src/features/auth/ui/sign-in/index.ts
@@ -1,1 +1,0 @@
-export { SignIn } from './SignIn'
